Remove emoji for common cases
the markdown mode is too emoji heavy for common cases,

| gem | before | after | compare |
|---|---|---|---|
| ddtrace | 0.18.1 | 0.18.2 | :x:  |
| i18n | 1.4.0 | 1.5.1 | :x:  |
| [nokogiri](https://github.com/sparklemotion/nokogiri) | 1.9.1 | 1.10.0 | :white_check_mark: [nokogiri@ 1.9.1...1.10.0](https://github.com/sparklemotion/nokogiri/compare/v1.9.1...v1.10.0) |

Remove the check mark from success, since it is already clear that the script
worked since it rendered the compare link. The green tick is confusing because
we haven't tested or checked anything, but it looks like the gem is approved?

Remove the red x from failures and replace with an error message, the red X
assumes that the developer has some understanding of what the red X means, but
without any context it's easy to assume the Gem has failed a check.  But it
could just be that the gem isn't hosted on Github.